Add pipeline evaluations and pipeline steps
Record evaluations of each predicate, and the steps in those evaluations.
Detect the log lines that signal completion/cache lookup of a predicate.
Note that cache lookups may not have tuple counts.
Update the log regexes to account for relations being printed as `name/arity@hash` instead of just `name`.