Add flame graph rendering with d3
Install d3 dependencies.
Add skipLib option to tsconfig.json files, otherwise the build fails on dependency errors within @types/d3-selection.
Have the top-level FlameGraph component communicate with the extension and update its state with the received data.
Have a FlameGraphDraw component within it that has the flamegraph data as its props.
This uses useEffect to perform d3 actions as side-effects, tracking props and the d3 container as dependencies.
It uses useRef to create a mutable reference to the d3 container.
See https://medium.com/@jeffbutsch/using-d3-in-react-with-hooks-4a6c61f1d102 for an example.